+++ lib/features/onboarding/screens/app_back_button.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:popcart/core/colors.dart';
+
+class AppBackButton extends StatelessWidget {
+  const AppBackButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      behavior: HitTestBehavior.opaque,
+      onTap: () {
+        Navigator.pop(context);
+      },
+      child: Container(
+        width: 48,
+        height: 48,
+        decoration: BoxDecoration(
+          color: AppColors.darkGrey,
+          borderRadius: BorderRadius.circular(16),
+        ),
+        child: const Icon(Icons.arrow_back_ios, color: AppColors.white),
+      ),
+    );
+  }
+}
