+++ lib/features/onboarding/screens/select_interests_screen.dart
@@ -1,55 +1,76 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:popcart/app/router_paths.dart';
 import 'package:popcart/core/colors.dart';
-import 'package:popcart/core/widgets/animated_widgets.dart';
+import 'package:popcart/core/utils.dart';
 import 'package:popcart/core/widgets/buttons.dart';
 import 'package:popcart/features/onboarding/cubits/interest_list/interest_list_cubit.dart';
 import 'package:popcart/features/onboarding/models/onboarding_models.dart';
-import 'package:popcart/features/onboarding/screens/enter_phone_number_screen.dart';
+import 'package:popcart/features/onboarding/screens/app_back_button.dart';
 import 'package:popcart/l10n/arb/app_localizations.dart';
 
-class SelectInterestsScreen extends HookWidget {
+class SelectInterestsScreen extends StatefulWidget {
   const SelectInterestsScreen({super.key});
 
   @override
+  State<SelectInterestsScreen> createState() => _SelectInterestsScreenState();
+}
+
+class _SelectInterestsScreenState extends State<SelectInterestsScreen> {
+  final List<ProductCategory> selectedInterests = [];
+
+  @override
+  void initState() {
+    super.initState();
+    context.read<InterestCubit>().getInterests();
+  }
+
+  Future<void> saveInterest() async {
+    final interests = context.read<InterestCubit>();
+    final selected = selectedInterests.map((e) => e.id).toList();
+    await interests.saveInterest(selected);
+  }
+
+  @override
   Widget build(BuildContext context) {
-    useEffect(
-      () {
-        context.read<InterestListCubit>().getInterests();
-        return null;
-      },
-      const [],
-    );
-    final interests = context.watch<InterestListCubit>();
-    final selectedInterests = useState(<ProductCategory>[]);
+    final interestCubit = context.watch<InterestCubit>();
     final l10n = AppLocalizations.of(context);
-    return Scaffold(
-      body: SafeArea(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              const AppBackButton(),
-              const SizedBox(height: 32),
-              Text(
-                l10n.select_interests,
-                style: const TextStyle(
-                  fontSize: 24,
-                  fontWeight: FontWeight.w700,
-                  color: AppColors.white,
+
+    return BlocListener<InterestCubit, InterestListState>(
+      listener: (context, state) {
+        state.whenOrNull(
+          saveInterestFailure: (message) {
+            context.showError(message);
+          },
+          saveInterestSuccess: () {
+            context.go(AppPath.authorizedUser.live.path);
+          },
+        );
+      },
+      child: Scaffold(
+        body: SafeArea(
+          child: SingleChildScrollView(
+            padding: const EdgeInsets.all(16),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const AppBackButton(),
+                const SizedBox(height: 32),
+                Text(
+                  l10n.select_interests,
+                  style: const TextStyle(
+                    fontSize: 24,
+                    fontWeight: FontWeight.w700,
+                    color: AppColors.white,
+                  ),
                 ),
-              ),
-              const SizedBox(height: 8),
-              Wrap(
-                children: interests.state.maybeWhen(
-                  loaded: (data) => data
+                const SizedBox(height: 15),
+                Wrap(
+                  children: interestCubit.interestsList
                       .map(
                         (e) => Padding(
-                          padding: const EdgeInsets.all(8),
+                          padding: const EdgeInsets.only(right: 10, bottom: 10),
                           child: ChoiceChip(
                             selectedColor: AppColors.orange,
                             showCheckmark: false,
@@ -57,55 +78,67 @@ class SelectInterestsScreen extends HookWidget {
                             shape: RoundedRectangleBorder(
                               borderRadius: BorderRadius.circular(100),
                               side: BorderSide(
-                                color: selectedInterests.value.contains(e)
+                                color: selectedInterests.contains(e)
                                     ? AppColors.orange
                                     : const Color(0xff50535b),
                               ),
                             ),
                             labelStyle: const TextStyle(
                               color: AppColors.white,
-                              fontWeight: FontWeight.w700,
-                              fontSize: 20,
+                              fontWeight: FontWeight.w400,
+                              fontSize: 14,
                             ),
-                            label: Text(
-                              e.name,
-                            ),
-                            selected: selectedInterests.value.contains(e),
+                            label: Text(e.name),
+                            selected: selectedInterests.contains(e),
                             onSelected: (value) {
-                              if (value) {
-                                selectedInterests.value = [
-                                  ...selectedInterests.value,
-                                  e,
-                                ];
-                              } else {
-                                selectedInterests.value = selectedInterests
-                                    .value
-                                    .where((element) => element != e)
-                                    .toList();
-                              }
+                              setState(() {
+                                if (value) {
+                                  selectedInterests.add(e);
+                                } else {
+                                  selectedInterests.remove(e);
+                                }
+                              });
                             },
                           ),
                         ),
-                      )
-                      .toList(),
-                  orElse: () => const [],
+                      ).toList(),
                 ),
-              ),
-              const SizedBox(height: 32),
-              IgnorePointer(
-                ignoring: selectedInterests.value.isEmpty,
-                child: AnimatedOpacity(
-                  opacity: selectedInterests.value.isEmpty ? 0 : 1,
-                  duration: const Duration(milliseconds: 300),
-                  child: BouncingEffect(
-                    onTap: () => context.go(AppPath.authorizedUser.live.path),
-                    child: CustomElevatedButton(
-                      text: l10n.next,
+                const SizedBox(height: 32),
+                if (interestCubit.interestsList.isNotEmpty)...{
+                  CustomElevatedButton(
+                    text: l10n.next,
+                    loading: context.watch<InterestCubit>().state.maybeWhen(
+                          orElse: () => false,
+                          loading: () => true,
+                        ),
+                    enabled: selectedInterests.isNotEmpty,
+                    onPressed: () {
+                      if (selectedInterests.isNotEmpty) {
+                        saveInterest();
+                      }
+                    },
+                  ),
+                  const SizedBox(height: 20),
+                  Center(
+                    child: GestureDetector(
+                      onTap: () {
+                        // final onboardingCubit = context.watch<OnboardingCubit>();
+                        context.go(AppPath.authorizedUser.live.path);
+                      },
+                      child: const Text(
+                        'Skip',
+                        style: TextStyle(
+                          fontSize: 16,
+                          fontWeight: FontWeight.w500,
+                          color: AppColors.orange,
+                        ),
+                      ),
                     ),
                   ),
-                ),
-              ),
-            ],
+                  const SizedBox(height: 32),
+                },
+              ],
+            ),
           ),
         ),
       ),
