+++ lib/features/onboarding/screens/video_splash_screen.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:popcart/app/router_paths.dart';
 import 'package:popcart/core/colors.dart';
-import 'package:popcart/core/widgets/animated_widgets.dart';
 import 'package:popcart/l10n/arb/app_localizations.dart';
 import 'package:video_player/video_player.dart';
 
@@ -19,11 +18,12 @@ class _VideoSplashScreenState extends State<VideoSplashScreen> {
   @override
   void initState() {
     super.initState();
-    _controller = VideoPlayerController.asset("assets/animations/popcart.mp4");
+    _controller = VideoPlayerController.asset("assets/animations/splash_animation.mp4");
     _controller
       ..addListener(() {
         setState(() {});
       })
+      ..setVolume(0)
       ..setLooping(true);
     _controller.initialize().then((_) => setState(() {}));
     _controller.play();
@@ -39,7 +39,7 @@ class _VideoSplashScreenState extends State<VideoSplashScreen> {
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context);
     return Scaffold(
-      backgroundColor: AppColors.orange,
+      backgroundColor: AppColors.white,
       body: Stack(
         children: [
           // Positioned.fill(
@@ -56,16 +56,13 @@ class _VideoSplashScreenState extends State<VideoSplashScreen> {
           Positioned.fill(
             child: _controller.value.isInitialized
                 ? FittedBox(
-                    fit: BoxFit.fitWidth,
+                    fit: BoxFit.fitHeight,
                     child: SizedBox(
                       width: _controller.value.size.width,
                       height: _controller.value.size.height,
-                      child: Container(
-                        padding: const EdgeInsets.only(left: 0.0, top: 15),
-                        child: AspectRatio(
-                            aspectRatio: _controller.value.aspectRatio,
-                            child: VideoPlayer(_controller)),
-                      ),
+                      child: AspectRatio(
+                          aspectRatio: _controller.value.aspectRatio,
+                          child: VideoPlayer(_controller)),
                     ),
                   )
                 : Container(),
@@ -76,30 +73,22 @@ class _VideoSplashScreenState extends State<VideoSplashScreen> {
               child: Column(
                 children: [
                   const Spacer(),
-                  // const SizedBox(
-                  //   width: double.infinity,
-                  // ),
-                  BouncingEffect(
-                    onTap: () {
-                      _controller.pause();
-                      context.push(AppPath.auth.path);
-                    },
-                    child: SizedBox(
-                      width: MediaQuery.of(context).size.width * 0.9,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          // context.go(AppPath.auth.path);
-                        },
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: const Color(0xffEDEAE9),
-                        ),
-                        child: Text(
-                          l10n.get_started,
-                          style: const TextStyle(
-                            color: Color(0xffF97316),
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                          ),
+                  SizedBox(
+                    width: MediaQuery.of(context).size.width * 0.9,
+                    child: ElevatedButton(
+                      onPressed: () {
+                        _controller.pause();
+                        context.push(AppPath.auth.path);
+                      },
+                      style: ElevatedButton.styleFrom(
+                        backgroundColor: const Color(0xffEDEAE9),
+                      ),
+                      child: Text(
+                        l10n.get_started,
+                        style: const TextStyle(
+                          color: Color(0xffF97316),
+                          fontSize: 16,
+                          fontWeight: FontWeight.w600,
                         ),
                       ),
                     ),
